% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_distance.R
\name{apply_scaling}
\alias{apply_scaling}
\title{Apply scaling to matching variables}
\usage{
apply_scaling(left_mat, right_mat, method = "standardize")
}
\value{
List with scaled left/right matrices and scaling parameters.
}
\description{
Apply scaling to matching variables
}
\keyword{internal}
