% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_blocks.R
\name{assign_blocks_group}
\alias{assign_blocks_group}
\title{Assign blocks based on grouping variable(s)}
\usage{
assign_blocks_group(left, right, block_by)
}
\value{
List with modified left/right data frames (with block_id) and n_blocks_initial.
}
\description{
Assign blocks based on grouping variable(s)
}
\keyword{internal}
