% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignr-package.R
\docType{package}
\name{assignr-package}
\alias{assignr-package}
\alias{_PACKAGE}
\title{assignr: Create Homework Assignments and Solutions using `rmarkdown`}
\description{
This package facilitates embedding solutions within an `rmarkdown`
document that can later be given to students.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/assignr}
  \item \url{http://thecoatlessprofessor.com/projects/assignr/}
  \item Report bugs at \url{https://github.com/coatless/assignr/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item David Dalpiaz \email{dalpiaz2@illinois.edu} [copyright holder]
}

}
