% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_join.R
\name{augment}
\alias{augment}
\title{Generic Augment Function}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{An object to augment}

\item{...}{Additional arguments passed to methods}
}
\value{
Augmented data (depends on method)
}
\description{
S3 generic for augmenting model results with original data.
}
