\name{bird_isotope}
\alias{bird_isotope}
\docType{data}
\title{
data: bird_isotope
}
\description{
Hydrogen and oxygen isotope of bird feather with known origin. Data is from Hobson et al., 2015.
}
\usage{data("bird_isotope")}
\format{
data.frame. Dimention is  105*6. 105 known origin stations. The columns are Species, Years, Latitude, Longitude, d2H and d18O from left to right.
}
\references{
Hobson, K.A. and Koehler, G., 2015. On the use of stable oxygen isotope (d18O) measurements for tracking avian movements in North America. Ecology and evolution, 5(3), pp.799-806.
}
\examples{
data(bird_isotope)
bird_isotope
}
\keyword{datasets}
