% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{couplr_messages}
\alias{couplr_messages}
\title{Couplr message helpers with emoji and humor}
\description{
Light, fun error/warning messages inspired by testthat, themed around
coupling and matching. Makes errors less intimidating and more memorable.
}
\keyword{internal}
