% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{err_invalid_param}
\alias{err_invalid_param}
\title{Invalid parameter error}
\usage{
err_invalid_param(param, value, expected)
}
\value{
No return value, throws an error.
}
\description{
Invalid parameter error
}
\keyword{internal}
