% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{err_missing_data}
\alias{err_missing_data}
\title{Missing data error}
\usage{
err_missing_data(dataset = "left")
}
\value{
No return value, throws an error.
}
\description{
Missing data error
}
\keyword{internal}
