% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignr.R
\name{get_example_filepath}
\alias{get_example_filepath}
\title{Retrieve example file path}
\usage{
get_example_filepath(x)
}
\arguments{
\item{x}{A `character` containing the name of the example Rmd.}
}
\value{
File path to the example file that ships with the package.
}
\description{
Obtains the file path for the example Rmd in the package.
}
\details{
The following example files ship with the package:
- hw00-main.Rmd
}
\examples{
get_example_filepath("hw00-main.Rmd")
}
