% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grader.R
\docType{package}
\name{grader-package}
\alias{grader}
\alias{grader-package}
\title{Create Homework Assignments, Solutions, and Rubrics using rmarkdown}
\description{
Grader enables the unibody design of homework solutions using rmarkdown.
}

