% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{info_low_match_rate}
\alias{info_low_match_rate}
\title{Low match rate info}
\usage{
info_low_match_rate(n_matched, n_left, pct)
}
\value{
No return value, called for side effects (issues a message or warning).
}
\description{
Low match rate info
}
\keyword{internal}
