% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_distance_cache.R
\name{is_distance_object}
\alias{is_distance_object}
\title{Check if Object is a Distance Object}
\usage{
is_distance_object(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical: TRUE if x is a distance_object
}
\description{
Check if Object is a Distance Object
}
\examples{
left <- data.frame(id = 1:3, x = c(1, 2, 3))
right <- data.frame(id = 4:6, x = c(1.1, 2.1, 3.1))
dist_obj <- compute_distances(left, right, vars = "x")
is_distance_object(dist_obj)  # TRUE
is_distance_object(list())    # FALSE

}
