\name{plover_range_BreedingSeason}

\alias{plover_range_BreedingSeason}

\docType{data}

\title{
Mountain plover breeding range}

\description{
Geographic range of North American mountain plover (\emph{Charadrius montanus}) during the breeding season.
}

\usage{data("plover_range_BreedingSeason")}

\format{SpatialPolygonsDataFrame}

\examples{
data("plover_range_BreedingSeason")
plot(plover_range_BreedingSeason)
}

\keyword{datasets}
