\name{plover_range_NonBreedingSeason}

\alias{plover_range_NonBreedingSeason}

\docType{data}

\title{Mountain plover non-breeding range}

\description{
Geographic range of North American mountain plover (\emph{Charadrius montanus}) during the non-breeding season.
}

\usage{data("plover_range_NonBreedingSeason")}

\format{SpatialPolygonsDataFrame}

\source{
\url{http://explorer.natureserve.org/servlet/NatureServe?searchName=Charadrius+montanus}
}

\references{
Ridgely, R. S. et. al (2003) Digital Distribution Maps of the Birds of the Western Hemisphere, version 1.0. NatureServe, Arlington, Virginia, USA.
}

\examples{
data("plover_range_NonBreedingSeason")
plot(plover_range_NonBreedingSeason)
}

\keyword{datasets}
