% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_distance_cache.R
\name{print.distance_object}
\alias{print.distance_object}
\title{Print Method for Distance Objects}
\usage{
\method{print}{distance_object}(x, ...)
}
\arguments{
\item{x}{A distance_object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Distance Objects
}
