% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_distance_cache.R
\name{summary.distance_object}
\alias{summary.distance_object}
\title{Summary Method for Distance Objects}
\usage{
\method{summary}{distance_object}(object, ...)
}
\arguments{
\item{object}{A distance_object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary Method for Distance Objects
}
