% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{warn_constant_distance}
\alias{warn_constant_distance}
\title{All distances identical warning}
\usage{
warn_constant_distance(value)
}
\value{
No return value, called for side effects (issues a warning).
}
\description{
All distances identical warning
}
\keyword{internal}
