% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{warn_constant_var}
\alias{warn_constant_var}
\title{Constant variable warning}
\usage{
warn_constant_var(var)
}
\value{
No return value, called for side effects (issues a warning).
}
\description{
Constant variable warning
}
\keyword{internal}
