% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{warn_many_forbidden}
\alias{warn_many_forbidden}
\title{Many forbidden pairs warning}
\usage{
warn_many_forbidden(pct_forbidden, n_valid, n_left)
}
\value{
No return value, called for side effects (issues a warning).
}
\description{
Many forbidden pairs warning
}
\keyword{internal}
