% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{warn_many_zeros}
\alias{warn_many_zeros}
\title{Too many zeros warning}
\usage{
warn_many_zeros(pct, n_zeros)
}
\value{
No return value, called for side effects (issues a warning).
}
\description{
Too many zeros warning
}
\keyword{internal}
