% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{warn_poor_quality}
\alias{warn_poor_quality}
\title{High distance matches warning}
\usage{
warn_poor_quality(pct_poor, threshold)
}
\value{
No return value, called for side effects (issues a warning).
}
\description{
High distance matches warning
}
\keyword{internal}
