% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignr-package.R
\docType{package}
\name{assignr-package}
\alias{assignr-package}
\alias{_PACKAGE}
\title{assignr: Create Homework Assignments and Solutions using 'RMarkdown'}
\description{
Writing homework assignments for students in the age of 'RMarkdown' 
    necessitates the creation of two separate documents -- assign.Rmd and soln.Rmd.
    The goal of assignr is to create one document main.Rmd that can be broken
    apart into the above two documents. Thus, there is no longer a need to copy
    and paste between the assign and the soln documents as all of the contents
    are together in one file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/assignr}
  \item \url{http://thecoatlessprofessor.com/projects/assignr/}
  \item Report bugs at \url{https://github.com/coatless/assignr/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item David Dalpiaz \email{dalpiaz2@illinois.edu} [copyright holder]
}

}
\keyword{internal}
