\name{assignR}

\alias{assignR}

\title{Tools for inferring geographic origin from isotopic data
}

\description{
Routines for rescaling isoscapes using known-origin tissue isotope data, assigning origin of unknown samples, and summarizing and assessing assignment results.
}

\author{
\strong{Maintainer}: Gabe Bowen \email{gabe.bowen@utah.edu}  
\strong{Authors}: Chao Ma, Gabe Bowen
}

\seealso{
\url{https://spatial-lab.github.io/assignR/}
}
