% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{couplr_warn}
\alias{couplr_warn}
\title{Warn with a fun, themed warning message}
\usage{
couplr_warn(..., call. = FALSE)
}
\description{
Warn with a fun, themed warning message
}
\keyword{internal}
