\name{plover_range_BreedingSeason}
\alias{plover_range_BreedingSeason}
\docType{data}
\title{
plover_range_BreedingSeason}
\description{
Mountain plover range of breeding season
}
\usage{data("plover_range_BreedingSeason")}
\format{
  The format is:
Formal class 'SpatialPolygonsDataFrame' with coordinate reference system of "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
}
\examples{
data(plover_range_BreedingSeason)
plot(plover_range_BreedingSeason)
}
\keyword{datasets}
