\name{plover_range_NonBreedingSeason}

\alias{plover_range_NonBreedingSeason}

\docType{data}

\title{Mountain plover non-breeding range}

\description{
Geographic range of North American mountain plover (\emph{Charadrius montanus}) during the non-breeding season.
}

\usage{data("plover_range_NonBreedingSeason")}

\format{SpatialPolygonsDataFrame}

\examples{
data("plover_range_NonBreedingSeason")
plot(plover_range_NonBreedingSeason)
}

\keyword{datasets}
