\name{plover_range_NonBreedingSeason}
\alias{plover_range_NonBreedingSeason}
\docType{data}
\title{
plover_range_NonBreedingSeason}
\description{
Mountain plover range for non-breeding season
}
\usage{data("plover_range_NonBreedingSeason")}
\format{
Formal class 'SpatialPolygonsDataFrame' with coordinate reference system of "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
}
\examples{
data("plover_range_NonBreedingSeason")
plot(plover_range_NonBreedingSeason)
}
\keyword{datasets}
