% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_messages.R
\name{success_good_balance}
\alias{success_good_balance}
\title{Perfect balance success message}
\usage{
success_good_balance(mean_std_diff)
}
\value{
No return value, called for side effects (issues a message).
}
\description{
Perfect balance success message
}
\keyword{internal}
